+++ src/data/models/transaction.py
@@ -2,7 +2,7 @@ import enum
 import uuid
 import datetime
 
-from sqlalchemy import DECIMAL, Enum, UUID, TIMESTAMP, Integer, ForeignKey
+from sqlalchemy import DECIMAL, Enum, UUID, TIMESTAMP, BigInteger, ForeignKey
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from src.data.models import Base
@@ -13,7 +13,7 @@ class TransactionModel(Base):
     __tablename__ = "transaction"
 
     id: Mapped[uuid.UUID] = mapped_column(UUID, primary_key=True)
-    user_id: Mapped[int] = mapped_column(Integer, ForeignKey('user.user_id', ondelete='CASCADE'))
+    user_id: Mapped[int] = mapped_column(BigInteger, ForeignKey('user.user_id', ondelete='CASCADE'))
     type: Mapped[enum.Enum] = mapped_column(Enum(TransactionType), nullable=False)
     cause: Mapped[str] = mapped_column(Enum(TransactionCause), nullable=False)
     time: Mapped[datetime.datetime] = mapped_column(TIMESTAMP(timezone=True), nullable=False, default=datetime.datetime.now(datetime.UTC))
